Errors of geoloc now go to stderr and added limitation for the vision_range equal to the number of solutions so that it doens't alloc too much memory.

+++ c_src/main.c
@@ -34,7 +34,7 @@ int main(int argc, char **argv){
     solution **sols = new_find_best_solutions(prob,
         pool_size, vision_range, &n_sols, &max_size_found);
     clock_t end = clock();
-    float seconds = (float)(end - start) / CLOCKS_PER_SEC;
+    float seconds = (float)(end - start) / (float)CLOCKS_PER_SEC;
     printf("Max size solution found: %d\n",max_size_found);
     printf("Search done in %f [s]!\n",seconds);
     // Print best solutions
